[ CMT-362-deposit-form ] refactor
