+++ src/scripts/front/components/Modal.js
