+++ src/scripts/front/rgbcode-authform.js
@@ -1,4 +1,4 @@
-import {Modal} from "./components/modal";
+import {Modal} from "./components/Modal";
 import {initTogglePass} from "./components/togglePassword";
 import {initFlagSelect} from "./components/flagSelect";
 import {initFormSubmit} from "./components/formSubmit";
